Replace tmux default-path option
Tmux dropped support for the default-path option - instead suggestion
to bind a key to create-window with the -c flag set. This works well
for creating new windows or panes with the same path as the current
window, but I prefer each session to have a default-path so no matter
what window I am in a new window behaves the same.

Add a function to
- Rename the session to get it out of the way
- Create a new session with the same name. Creating the session sets
  the directory for new windows to $PWD
- Move all windows over to the new session and switch to it

Works best with set-option -g renumber-windows to keep window indexes
the same in the new session