Add mappings to switch a method to/from =>
Quick and dirty first pass. These mappings clobber search and depend on
the cursor being within the line containing '=>' or 'return' and may
have bad side effects if they aren't.

In the future I'd like to try to write a proper toggle so I can use the
same key binding for both directions and bail out if the cursor isn't
inside a toggleable method