Support projects in a subdirectory
When working on the SDK the "root" of the project has some gclient
config but is not the root of the git directory. I always want to be in
the `sdk/` subdirectory in that case.

Detect when the root of a project is not a git repo, but does have a
single non-hidden subdirectory which is a git repo, and move there
instead.