Filter some zsh history
There is no convenient way to remove an entry from the `$HISTFILE` by
command number or similar, so use some sed magic to delete the commands
that have a matching timestamp. This relies on the `EXTENDED_HISTORY`
option.

Delete commands which:
- Have exit code 64 which corresponds to bad usage.
- Have exit code 127 which is set by the shell for command not found.
- Are `echo` which multiple lines since those are noisy and not useful
  to repeat.
- Are `git` and exit with 129 which is the code uses for incorrect usage
  of inner commands. A missing command will still exit with 1.
  https://github.com/git/git/blob/53f9a3e157dbbc901a02ac2c73346d375e24978c/Documentation/technical/api-error-handling.txt#L10-L12

The history will remain locally, but won't be persisted or shared with
other open shells.

There is an edge case where a line in the middle of a multiline command
which happens to have a timestamp in the format of the history file at
the beginning of the line could get deleted, but this is incredibly
unlikely.