Autoclose utility windows on quit
When closing a window, if it was the last "editor" window, close
quickfix and preview windows as well. This works well for tab pages and
for the entire process on the last tab. Stops the annoying case of
having to also close windows I don't care about to quit entirely.

Move the window local location list closing into a new plugin called
"autoclose" and add functionality to check whether there is only 1
normal window left before closing all utility windows.