Add late feature branching to git-start
Handle the case where I forget to start a feature branch before making
the first commit. In the special case of being on the master branch
which is tracking a remote, reset master back to the same state as the
remote and avoid leaving commits that could accidentally get pushed to
master.

- Move from a simple alias to a bash script and add a man page.
- Before starting the branch check for the case of starting on "master"
  and having commits that aren't pushed to a remote. If we are record
  the fact that we want to reset the branch.
- After starting the new branch, pop back and reset with --hard if
  necessary.