Stay in visual mode with repeated 'v'
When in visual mode hitting any of 'v', 'V', or "<c-v>" would switch to
the corresponding visual mode, however hitting the key for the current
mode would go back to normal mode. Add mappings to make hitting the key
for the current mode a no-op so that repeated presses, or accidental
presses if I was already in that mode, are ignored.