Download Dart SDKs to separate directories
Instead of downloading the SDK fresh each time it's switch from dev to
stable, download the SDKs to unique directories by version and symlink
for the overall SDK.

When a specific version is not specified or is "latest", query the
`VERSION` URL to find what the latest version is. If the version is
already downloaded skip the download, otherwise unzip then move into
`~/.dart_sdks/$VERSION`.

Requires `jq` on the command line to parse the JSON response.