Add projects in session_finder
Include projects in the list of sessions. Add a postfix `P` or `*` to
distinguish projects from sessions. If a session already exists for a
project it is listed as a session, not a project. Move the directory
finding logic from the `project` command to `session_finder` for use
when opening a project without an existing session.

Remove the `project` script and bind-key.

- Use `mapfile` instead of the longer `read` approach.
- Use `tmux has-session` over checking the array in bash.
- When there are no sessions, add default "0".
- Replace `.` for projects with `_` to make a valid tmux session name.
- `cd` to home directory for session "0". Remove from tmux bind-key.
- Add `set -e` to fail fast.
- Fix `shellcheck` warnings.
- Remove compdef for removed project sctipt.