Allow more renames in git
When working in the Dart SDK we can occasionally hit very large merges
that don't play well with the current prompt. Increasing the
`renameLimit` helps to avoid some output on stdout, but performance is
still so slow that it makes it hard to get a command to run with the
prompt interrupting every second. Descrease the update interval to 10
seconds to compensate.