Drop some install steps
python-dev doesn't seem necessary anymore, I think it had been a
prerequisite for something else but the original commit message doesn't
indicate what, and nothing seems to fail without it now.

rvm was only necessary for jekyll, which I have not touched in a while.

Drop vim in favor for manually compiling to get the latest.