Use the main-vertical layout for tmux_side_pane
Use a larger main pane for very wide screens. On a screen that is at
least 400 characters wide I can fit 3 vim splits in the main pane, so
aim for ~1/4 window width on a wide screen, and ~1/3 window width on a
smaller screen. Put some buffer in to avoid an overly narrow right pane.

Use the `main-pane-width` option and the `main-vertical` layout to have
tmux resolve the full layout rather than hardcode the approximate layout
and hope that it scales well.

This does lose the option to have the secondary pane on the left, since
`main-vertical` layout is always main pane on the left, but I rarely
used this. Instead bind that to select the `even-horizontal` layout
which is more useful when there isn't a "main" pane.