Don't de-indent a list nested in a list
Previously a list like:

```
some_key:
- another_key:
  - listvalue
```

Would cause the nested list to de-indent and change the meaning. Once a
new key is started in a list value, it's own list should be indented to
the level of the key, not the level of the outer list.

Trim whitespace before checking the lines to avoid extra trickiness
around allowing whitespace but not a `-` character. Don't remove
indentation when the previous line was both a list entry and a key
without a value.