Add a search command in vim
Opens an FZF popup in phony mode so the query is fed directly to `ag`
instead of used as a fuzzy filter on the results. This is easier to
quickly move around and iterate on the query than using the quickfix
list.

When more than one result is selected it will populate the quickfix
list, and all results can be selected with `<c-a>`.

Change the grep operator to fill out the initial query where it can then
be used immediately, or refined. Add a shortcut to open with a blank
query. The lowercase versions use the local working directory, while the
capital versions use the global working directory. Take back the
lowercase mapping from it's old use of creating a scratch buffer by
moving that functionality to the shortcut to edit notes which previously
would error when outside a project. Now either the projects notes will
be opened, or when not in a project a scratch buffer is used instead.