Fix manpath appending
The `$MANPATH` environment variable does not initially exist, so the
current pattern causes only the local man files to be accessible.
Instead set the variable directly and end with a `:` to postfix the
manpath which is discovered automatically.

Eventually it would be good to figure out how to get these put in a
place where they can also be discovered automatically.