Adjust vim colors for terminal mode
When using termguicolors the background of the terminal will fall back
to black (based on `background`) instead of the terminal's color. Assign
the same background and foreground color in `NORMAL` highlight group,
which is used for the terminal, as I was using in one terminal. I will
need to adjust all terminals to use the same.

Drop the EndOfBuffer config since it seems to have no effect anymore.

Link `StatusLineTerm` to `StatusLine` to avoid obnoxious green
statusline for terminal windows.

Add a status mode marker for `t` mode. Use the same color as insert mode
since the keystrokes would go to the buffer.