Limit fuzzy file find to current package
Attempt to reduce the dependency on the local working directory for
`<leader>lf` to do the most useful thing. List files in the local
package by default, and fallback to the local working directory if the
package is unknown.

It may also be worthwhile to add specific keybindings for finding across
Git, or finding across the local directory. `lF` is still the global
directory which should work well for project layouts with individual
repo copies.