Find todos against default branch
Before pushing for a PR there will always be a diff against the upstream
branch for any new code, because I always set the upstream branch with
`git start`. But after pushing for a PR or, more relevantly, a
long-lived feature development branch there may frequently be no diff
against the upstream. After pushing to a remove it is more useful to
compare against the default branch.