Wrap full words with \b in search operator
Makes some searching for some of the most common operators, like `iw`,
work more like expected because they will match whole words only.

Check whether the chosen range matches a search for the exact content of
that range with `\<` and `\>` at the ends to indicate that the ends are
also the first and last characters of a word. The `\%V` token limits the
search to the last visual range, which will match the selected range
since the yank uses visual mode.

When the selection stats on the first character of a word, and ends on
the last character of a word, wrap the query with `\b` to also enforce
word boundaries. This may not behave exactly as expected when searching
for a query with multiple words, but it seems to help in more cases than
it hurts.