Drop TmuxNavigatPrevious bind
I don't use it, `<c-w><c-w>` accomplishes the same within vim and it's
not necessary to flip between tmux pane and vim, especially with more
terminals staying within vim. Avoid the built in mappings, and map only
the required keys.
Allow the `<c-\>` to be used for the default behavior and force normal
mode when followed by `<c-n>`.