Add commands to edit files in surrounding vim
Add an alias `e` to open vim or drop a file into the already running vim
depending on whether the shell is running in vim.

Set `EDITOR` and `VISUAL` to vim or a script that opens the file and
switches the terminal into normal mode to avoid interacting with it and
giving it output before the buffer is closed. Open the file and set
`bufhidden=delete` so that `BufUnload` is a reasonable signal that edits
to the file are complete. Add a `Tapi_` vim function to open a split or
switch to an existing buffer with the passed file. When the buffer is
unloaded after edits are done, send a message back to the terminal which
is waiting for any input to consider the edit complete.